Refactor tournaments page flow and update shared assets
- replace the tournaments page with an authenticated, responsive implementation and inline Search/Filter components\n- remove legacy tournament UI components now embedded in the page module\n- wire the sidebar tournament link and fix case-sensitive sidebar import usage\n- normalize tournament icon SVGs to use currentColor/currentStroke for theming
